commit 6

+++ src/main/java/ru/maks/kurs/dao/StudentDao.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface StudentDao extends JpaRepository<Student, Long> {
 	Optional<Student> findByFirstName(String name);
-	Optional<Student> findStudentByClassNumber(Long class_number);
+	Optional<Student> findStudentByClassNumber(Integer class_number);
 	Optional<Student> findByDateOfContractAfter(String dateOfContract);
 	Optional<Student> findByFirstNameAndLastName(String name);
